requirements.txt
PyQt5
requests
PyGithub
ollama
python-dotenv